requirements.txt
Django==3.0.7
django-mptt==0.11.0
djangorestframework==3.11.0
rest-framework-generic-relations==2.0.0
munch==2.5.0
Cerberus==1.3.2
coloredlogs==14.0
PyGObject==3.36.1
WeasyPrint==51
python-gnupg==0.4.6
PyYAML==5.3.1
PyPDF2==1.26.0
